# aurigraph-av10-7/aurigraph-v11-standalone/04-documentation/openapi.yaml
openapi: 3.0.3
info:
  title: Aurigraph V11 Platform API
  version: 11.0.0
  description: |
    **Aurigraph V11** is a high-performance blockchain platform built on Java/Quarkus/GraalVM,
    designed to achieve 2M+ TPS with quantum-resistant cryptography and AI-driven optimization.

    ## Key Features
    - **HyperRAFT++ Consensus**: Advanced consensus algorithm with AI optimization
    - **Quantum-Resistant Cryptography**: CRYSTALS-Dilithium/Kyber (NIST Level 5)
    - **Cross-Chain Bridge**: Interoperability with Ethereum, Solana, and other chains
    - **AI/ML Optimization**: Machine learning-driven performance tuning
    - **Real-World Asset Tokenization**: Enterprise-grade asset digitization
    - **Ultra-High Performance**: Target 2M+ TPS with <100ms finality

    ## Authentication
    All authenticated endpoints require a Bearer token:
    ```
    Authorization: Bearer <your-jwt-token>
    ```

    ## Rate Limiting
    - API endpoints: 100 requests/second
    - Admin endpoints: 10 requests/second
    - Auth endpoints: 5 requests/minute

    ## Base URL
    - Production: https://dlt.aurigraph.io
    - Development: http://localhost:9003
  contact:
    name: Aurigraph DLT Team
    email: support@aurigraph.io
    url: https://aurigraph.io
  license:
    name: Proprietary
    url: https://aurigraph.io/license
servers:
  - url: https://dlt.aurigraph.io
    description: Production server
  - url: http://localhost:9003
    description: Local development server
...
paths:
...
  /api/v11/performance:
    get:
      tags:
        - Core Platform
      summary: Performance test
      description: Run performance benchmark test
      operationId: performanceTest
      parameters:
        - name: iterations
          in: query
          description: Number of transactions to process
          required: false
          schema:
            type: integer
            default: 100000
            maximum: 500000
        - name: threads
          in: query
          description: Number of parallel threads
          required: false
          schema:
            type: integer
            default: 1
            maximum: 64
      responses:
        '200':
          description: Performance test completed successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PerformanceStats'
        '429':
          description: Rate limit exceeded
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
...
  /api/v11/blockchain/transactions:
    post:
      tags:
        - Blockchain
      summary: Process transaction
      description: Submit a transaction for processing
      operationId: processTransaction
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/TransactionRequest'
            example:
              transactionId: tx_12345
              amount: 100.50
      responses:
        '201':
          description: Transaction processed successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TransactionResponse'
        '400':
          description: Invalid transaction data
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '503':
          description: Service temporarily unavailable

    get:
      tags:
        - Blockchain
      summary: Get transactions
      description: Retrieve list of recent transactions
      operationId: getTransactions
      parameters:
        - name: limit
          in: query
          description: Maximum number of transactions to return
          schema:
            type: integer
            default: 10
            maximum: 100
        - name: offset
          in: query
          description: Offset for pagination
          schema:
            type: integer
            default: 0
      responses:
        '200':
          description: Transactions retrieved successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TransactionList'
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: JWT token for authenticated endpoints
  schemas:
...
    PerformanceStats:
      type: object
      properties:
        iterations:
          type: integer
          example: 100000
        durationMs:
          type: number
          format: double
          example: 54.32
        transactionsPerSecond:
          type: number
          format: double
          example: 1840000.0
        nsPerTransaction:
          type: number
          format: double
          example: 543.2
        optimizations:
          type: string
          example: Java/Quarkus + Virtual Threads
        threadCount:
          type: integer
          example: 8
        targetTPS:
          type: integer
          format: int64
          example: 2000000
        targetAchieved:
          type: boolean
          example: false
...
    TransactionRequest:
      type: object
      required:
        - transactionId
        - amount
      properties:
        transactionId:
          type: string
          example: tx_12345
        amount:
          type: number
          format: double
          example: 100.50
    TransactionResponse:
      type: object
      properties:
        transactionId:
          type: string
        status:
          type: string
          enum: [PROCESSED, PENDING, FAILED]
        amount:
          type: number
          format: double
        timestamp:
          type: integer
          format: int64
        message:
          type: string
...
    TransactionList:
      type: object
      properties:
        transactions:
          type: array
          items:
            type: object
            properties:
              hash:
                type: string
              from:
                type: string
              to:
                type: string
              value:
                type: string
              fee:
                type: string
              status:
                type: string
              timestamp:
                type: integer
                format: int64
              blockHeight:
                type: integer
                format: int64
        total:
          type: integer
          format: int64
        limit:
          type: integer
        offset:
          type: integer
...
    Error:
      type: object
      properties:
        error:
          type: string
          example: Invalid request
        message:
          type: string
          example: Detailed error message
        code:
          type: integer
          example: 400
        timestamp:
          type: integer
          format: int64